Convert app from SPA mode to server mode
I didn't realise that the PhaserJS template was in SPA mode, which
breaks Remix's default Form functionality, requiring that fetch
be used for all backend requests. I want to be able to use forms
for the web app CRUD functionality, so I removed the SPA mode
config and replaced it with the default Remix config.