Wrap Remix Link in a styled component
The move away from SPA mode in the Remix/Vite config broke wrapping
a Remix Link in a Chakra UI Link (or passing it to Emotion's styled
constructor). So, I'm wrapping the Link component in a styled div
that adds button-like styles to the rendered link. Probably better
to just put this in a *.css file at this point, but it's fine for
now.