Add page for creating terrain to add to game
I made this a separate page/form to break down the game creation
process into more-manageable steps. Also, terrain could easily
be represented by JSON objects in the games table, but it's not
a bad idea to just make them records in their own table for a bit
of future-proofing.