Upgrade React to v19
We have to use a "canary" version, because the current stable
version is 18.3.1, but React 18 totally broke JS frameworks with
SSR and injected JS/CSS in the head (e.g. Remix, Next). The app
worked okay, but Cypress kept raising errors about hydrated DOM
not matching the DOM rendered by the server.

A comment on an issue related to this suggested that this version
of React fixed the hydration mismatch error, and it did.

I got a bunch of incompatible version warnings when I upgraded
React, but the app runs fine, so I'll hold off on updating other
packages until v19 is officially released.