Get e2e tests running in docker-compose
I had to include a bunch of packages in the vite config, because
there's a known issue where vite keeps optimising dependencies
in CI/Docker, which can cause tests to fail due to the hot reloading
that's triggered. This seems to fix it, and I'll just have to add
more packages to the list as the flake pops up.