Refactor basic Chakra components into components folder
I want to isolate Chakra UI's components a bit from the rest of
the app to make it easier to migrate to another component library
if I choose to do so in the future. I won't go 100% on this,
because that feels like overkill, but at least the most-commonly
used components won't be duplicated throughout the app, with
only one place to replace the Chakra version with something else.